Moved lexer.h to satisfy cpplint
