+++ rcl/test/CMakeLists.txt
@@ -72,7 +72,6 @@ function(test_target_function)
     ENV ${extra_test_env}
     APPEND_LIBRARY_DIRS ${extra_lib_dirs}
     LIBRARIES ${PROJECT_NAME} ${extra_test_libraries}
-    INCLUDE_DIRS ${CMAKE_SOURCE_DIR}/src
     AMENT_DEPENDENCIES ${rmw_implementation}
   )
